pyproject.toml
[tool.poetry]
name = "websoket-fastapi"
version = "0.1.0"
description = ""
authors = ["Vladislav-Shi <vlad.shirobockow@yandex.ru>"]
readme = "README.md"
packages = [{include = "websoket_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.86.0"
pydantic = "^1.10.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
python-dotenv = "^0.21.0"
uvicorn = "^0.19.0"
aerich = "^0.7.1"
pytest-asyncio = "^0.20.2"
httpx = "^0.18.2"


[tool.poetry.group.dev.dependencies]
requests = "^2.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
